Refactor code to create common component for session insights tab
