+++ packages/extension/src/view/devtools/pages/privacyProtection/ipProtection/probabilisticRevealTokens/sessionInsights.tsx
@@ -17,13 +17,12 @@
  * External dependencies.
  */
 import React from 'react';
-import { CirclePieChart } from '@google-psat/design-system';
 
 /**
  * Internal dependencies.
  */
 import { useProbabilisticRevealTokens } from '../../../../stateProviders';
-import classnames from 'classnames';
+import InsightsStats from '../../mdlCommon/insightsStats';
 
 const SessionInsights = () => {
   const { statistics } = useProbabilisticRevealTokens(({ state }) => ({
@@ -57,43 +56,7 @@ const SessionInsights = () => {
     },
   ];
 
-  return (
-    <div>
-      {stats.map(
-        ({ title, centerCount, color, onClick, data, tooltipText }, index) => (
-          <button
-            key={index}
-            className={classnames('group text-center w-20 p-2 h-full', {
-              'active:opacity-50 hover:scale-95 transition-all duration-300 ease-in-out':
-                onClick,
-            })}
-            style={{ cursor: !onClick ? 'default' : 'pointer' }}
-            onClick={() => {
-              onClick?.();
-            }}
-          >
-            <CirclePieChart
-              key={title}
-              title={title}
-              centerCount={centerCount}
-              data={data ?? [{ count: 10, color: color ?? '' }]}
-              infoIconClassName="absolute -right-3"
-              bottomTitleExtraClasses="min-w-[150px]"
-              tooltipText={tooltipText}
-              centerTitleExtraClasses={classnames({
-                'group-hover:scale-125 transition-all duration-300 ease-in-out':
-                  onClick,
-              })}
-              pieChartExtraClasses={classnames({
-                'group-hover:scale-[1.15] transition-all duration-200 ease-in-out group-hover:bg-[#f3f3f3] dark:group-hover:bg-[#191919] rounded-full':
-                  onClick,
-              })}
-            />
-          </button>
-        )
-      )}
-    </div>
-  );
+  return <InsightsStats stats={stats} />;
 };
 
 export default SessionInsights;
